core: make Genesis::records and records_file private (#7426)
Genesis::records are not guaranteed to be set since the records may
reside in records file instead.  Code accessing that field directly
(without taking records_file field into consideration as well) is
technically incorrect and violates Genesis’ encapsulation.

Make the records an records_file private to force all users of the
class to uses provided methods rather than accessing the fields
directly.  To accommodate all usage, introduce records_len and
force_read_records methods.

Note that this isn’t the state we’d like the struct to be in.
Ideally, we’d have an enum which either holds the records in a vector
or path to the file with the records.  This commit is a step into that
direction.