store: make StoreOpener::open return Result rather than panicking (#7461)
Having a Result allows for the error to be propagated to the user in
nicer form than in a panic message.  This will also help in the future
expansion of the StoreOpener::open method to return different kinds of
failures which may be handled by the caller in different ways.