store: avoid Vec conversion when getting data from RocksDB (#7494)
Currently, when data is read from RocksDB it’s converted from
RocksDB’s internal type (specifically DBPinnableSlice) to a vector.

The problem is that we often end up discarding this vector just as
soon as it’s created.  For example, when deserialising objects using
get_ser method an owned vector is unnecessary.  And when accessing
DBCol::State the data is converted into an Arc which forces another
allocation and memory copy.

Introduce DBBytes type which is what Database::get_raw_bytes and
Database::get_with_rc_stripped methods return.  For RocksDB the object
holds RocksDB specific DBPinnableSlice which than caller can treat as
a slice or convert to a vector or an arc as desired.

Furthermore, change Store::get to return data in this form.
Furthermore, for convenience, introduce Store::get_vec which does
conversion to a vector.