refactored PeerActor::new (#7503)
This is part of a preparation to support TIER1 peers.
I've moved as much logic as possible inside the constructor, so that the number of arguments is minimal. I've also added a PeerActor::Config which is passed to the constructor, because PeerActor::new cannot return an error (due to Actix API constraints). Also moved accepting new connections from PeerManagerActor to NetworkState.