store: allow configuring RocksDB with support for subset of columns (#7544)
With cold storage approaching, it’s gonig to be necessary to be able to
open RocksDB instance with only a subset of defined columns.  Change how
RocksDB deals with column family handles.  Specifically, get_cf_handles
now accepts an iterator over DBCol values as argument and cf_handle will
return an error if trying to get a handle for a column that wasn’t set
up.