store: propagate error when deleting the snapshot (#7594)
If Snapshot::remove failed to delete the snapshot it would log the
error but wouldn’t propagate it letting the node continue to work. The
reasoning was that the database has migrated so there’s no harm just
going on.

However, it’s possible that operator won’t notice errors in the log
and leave the snapshot around.  This will waste space but also cause
next migration to fail.  What’s worse, that failure will recommend
rolling back to the existing snapshot.  This may result in database
being rolled back to state from weeks or months ago.

Change this behaviour by making Snapshot::remove propagate the error
and opening of the database failing if the removal fails.