removed deprecated RoutedMessageBody::PartialEncodedChunk (#7610)
PartialEncodedChunk was deprecated in favor of VersionedPartialEncodedChunk for 2 years now: https://github.com/near/nearcore/pull/3463
It is confusing, in particular we have configured PartialEncodedChunk to be sent out 3x (to make it delivered reliably) while
we should have done that for VersionedPartialEncodedChunk (which is still sent out just 1x).

There is a codepath which encodes V1 PartialEncodedChunk into this specific RoutedMessagebody variant, however while decoding this distinction is lost. Technically this PR should be split into 2 separate rollouts (first: stop using the deprecated variant to send messages; second: stop accepting the deprecated variant). The situation gets even more complicated, because network protocol is supposed to be backward compatible for 2 versions (so, 3 rollouts), and since this is a routed message you can imagine a sequence of nodes where node in version N sends a message to node in version N+4 via node in version N+2 (and theoretically no amount of rollouts helps then, because the intermediate nodes cannot translate the message to a newer format, because it is signed).

In spite of all of that AFAICT it is safe to merge this PR in one go, because VersionedPartialEncodedChunk is used only to send new chunks from chunk producer to approvers. Therefore it is used only with V2 PartialEncodedChunk for quite some time now.