core: replace ShardChunkHeader::version_range by valid_for (#7630)
Replace ShardChunkHeader::version_range and
PartialEncodedChunk::version_range methods with valid_for method which
is simpler interface for the call site and which allows us to get rid
of custom ProtocolVersionRange type.

Note that, yes, PartialEncodedChunk::valid_for is actually unused.
I’m leaving it in though.  Perhaps it’s worth keeping the record of
when each version of the PartialEncodedChunk was introduced?