store: refactor handling of database migration (#7584)
Move logic of driving the database migration to StoreOpener.  This is
done by introducing StoreMigrator trait which is passed when creating
StoreOpener through a new NodeStorage::opener_with_migrator method.
Object implementing that trait contains code for performing migrations
one level at a time and StoreOpener is what is now checking the
database version and looping through all the migrations calling the
migrator.

The change is motivated by future plans of adding cold database to
NodeStorage.  With cold database, versions of two databases will need
to be checked and it’s cleanest to encapsulate all that logic in
StoreOpener.  This makes rest of the code pretty much unaware of how
versions are stored and only dedicated migration code needs to deal
with performing single-version migrations.