fix: compare head heights before updating flat storage (#7724)
In `tests::client::process_blocks::test_query_final_state`, there is a chain with two forks:

```
///                  /-------> 6
/// 1 -> 2 -> 3 -> 4 
///                  \---> 5
```
When (5) is postprocessed, flat head moves from (2) to (3).
When then (6) is postprocessed, the last final block it sees is (2). It attempts to move flat head from (3) to (2) and the node crashes.

Also we can't simply take final head which always moves forward, due to catchups. So we add comparison of old and new heights.

## Testing

https://buildkite.com/nearprotocol/nearcore-flat-state/builds/62#01838983-7f4d-4bd1-9eea-a376da2e0a62