fix: make non-integration tests work with flat storage (#7816)
This PR fixes remaining non-integration tests for flat storage. Two tests needed to be disabled because flat storage doesn't support sync and sharding upgrade. Also, NightshadeRuntime tests were the most problematic here.

We need to initialize FlatStorage manually because creation naturally happens on Chain, but runtime tests doesn't use it and operate on mock chain instead which is just a sequence of blocks with hashes `hash(&[height])`. Also we need to update FlatStorage with new deltas, because this happens on Chain as well but in more complex fashion. Here we can just add new delta and skip logic for updating flat head.

## Testing

* Existing tests
* https://buildkite.com/nearprotocol/nearcore-flat-state/builds/74
* https://nayduck.near.org/#/run/2716