chain: reuse more types in JSON RPC fuzzer (#7782)
Rather than redefining a lot of the types used in JSON RPC requests,
reuse existing type definitions.  This has the effect of making the
fuzzer more in sync with the actual code that it’s fuzzing.

Some types are left as is because they would require implementing
Arbitrary for crypto-related types (public key, signature etc.) and at
this time I’m not sure what’s the best approach to do it.

While at it, rewrite implementation of Base64String serialisation to
reuse existing base64_format code.