feat: estimate `ed25519` base and bytes separately (#7980)
The old estimation just divided the base estimation by the number of
bytes to receive the cost per byte. Now the base is subtracted first and
the remainder is divided by the number of bytes.

Exact estimations will be done and added later. But icount estimations
show that the per byte cost can be reduced from the placeholder value
423 Mgas to 12 Mgas (safety multiplier of 3 included).

The base value is still a bit tricky to say. `icount` estimation gives
me 90 Ggas, so with a safety multiplier it would be 270 Ggas.
The current placeholder of 40 Ggas is quite a bit lower but it was
roughly what we expected based on
[external benchmarks](https://github.com/dalek-cryptography/ed25519-dalek#benchmarks).
It could be that qemu based estimation does not work well here because
AVX instructions are used.