Introduce ‘stdx’ crate with split array polyfills (#8087)
This consolidates all the places where array splitting happens into
a single module.  In the future, once split_array_ref feature gets
stabilised, we’ll be able to get rid of those polyfills but for now
it’s nice to have them all in one place.

The new `stdx` crate also provides a place where other common utility
features can be put so that the same function doesn’t need to be
implemented in multiple crates or put in a crate it doesn’t fit in and
made public just because some other crate depends on it.

Crucially, the crate must not depend on any crates from the workspace.