fix(mirror): fix bug in handling of access keys that have been deleted (#8104)
right now we're just storing the source and target nonces for added keys, but then we don't update anything if the key is deleted and readded. So to fix this, we'll store only the target chain nonce on disk, along with the hashes of txs and receipts that might update it in the future.

A big downside to this is that the logic around handling nonces is much more complicated and probably more bug prone. But doing this will make it easier to get around the problem of access keys added by FunctionCalls by just sending extra transactions.

This PR also adds traffic to the corresponding pytest that fails before this change, but passes after (except for a few mishaps where the AddKey tx hasn't made it on chain by the time we want to use it)