Made type of ArcMutex::update more strict. (#8126)
The semantics of `ArcMutex::update` was misleading in case the result type was `Result<_,_>`, as one could think that returning an error would cancel the update (which was not the case: any mutations done before returning the error were commited). With the new API, ArcMutex::update takes a state monad as an argument (a function which takes old state and returns a new state with an optional result) which makes it obvious that the returned state is actually applied. I've also added a separate ArcMutex::try_update method which allows to cancel the update.