runtime: introduce PublicKeyBuffer to avoid memory allocation (#8277)
Currently when public key is read it’s unnecessarily copied into
a temporary vector.  The copy is necessary because otherwise public
key data would hold a shared reference on self.registers which
prevents any methods which take exclusive reference to self to be
called.

However, if we decode the public key immediately after reading it, we
can deal with sized objects which can be stored on stack.  This is
perfect except because we cannot change when errors are reported we
have to postpone returning of the deserialisation errors.

Introduce PublicKeyBuffer which facilitates that.  It hides from the
users the fact that deserialisation happens immediately as soon as the
data is read providing an interface which looks as if deserialisation
was happening separately from reading.

An alternative would be to use a fixed-size 65-byte array and separate
reading the data from deserialisation just like it’s happening
now. A minor downside of that is hard-coding the 65-byte size.  We’d
probably want PublicKey::MAX_LEN but even then having such a constant
doesn’t seem too clean of a solution to me.