chore: tokio patch version update to 1.18.4 (#8309)
replaces #8304

dependabot made us aware of bugfixes and wanted to update to 1.20.3

Instead I propose we update only the patch version and keep major/minor.
1.18.x is a LTS version and it is what we are currently using.
1.18.4 contains the same fixes the bot wants.
Updating further should be done as a conscious decision.

`version = "~1.18"` means we want a fixed 1.18 minor version with
the newest patch version.

The old `"1.16.1"` is equivalent to `^1.16.1"` but in the lock file it
was fixed to `"1.18.2"`. Since we check in the lock file and only update
versions occasionally and consciously, I suggest we use `~` which is
less confusing.