refactor: Migrate parameter configs to YAML (#8310)
This is a refactoring PR to enable https://github.com/near/nearcore/issues/8264. For full motivation see https://github.com/near/nearcore/issues/8264#issuecomment-1373628858

This PR only changes the disk format and parsing logic but doesn't change the in-memory representation, so it should not lead to any visible behaviour changes.

The next step in this refactoring would be to investigate getting rid of `serde_json::Value` in favor of `serde_yaml::Value` or even better a specialized struct with much more restricted types to catch invalid config errors earlier.