Fix Nayduck test repro_2916.py. (#8454)
This is due to a change in the code that, when a request could only find some of the parts or receipts that were requested, instead of not responding at all, would now respond with whatever was found. This caused the nayduck test to fail because it was checking the existence of a response. We now check exactly what the response was, and make the test much clearer about what it's asserting.

Also, I noticed that the test framework assumes that the cargo target dir is "./target", and added support to honor the CARGO_TARGET_DIR env var if specified.