abort_on_panic in tests (#8439)
Made some of the tests abort on panic. We started using "cargo nextest" instead of "cargo test" in CI, so now every test is executed in a separate process and aborting a test won't kill the test harness. This will potentially reveal some flaky tests, which behave in an unexpected way (and panic) but report success anyway. In a long term it is expected to make debugging tests easier, since they will terminate as soon as first panic occurs. Enabling this behavior is per test, since we have some tests which actually expect a panic.