feat: reject delegate actions on older protocol (#8477)
Without this change, a binary that has support for meta transactions
would accept them even before the protocol upgrade has
taken effect.

The solution is to add a transaction verification error.
For that, I had to make the action validation protocol version dependent,
which it hasn't been so far.
Changes made to the validation code should be
non-observable.

Note: This does not solve potential issues on the network layer
during the upgrade. It will have to be solved separately.

### Test strategy
----
One test for both the behaviors, before and after the protocol upgrade.