feat: allow creating cold store when hot store exists (#8510)
Currently the opener expects both stores to exist or not exist. The case when hot store exists and should be openened and cold store doesn't exist and should be created is not supported. 

This PR addresses this issue - allowing opening hot and creating cold - as well as refactors some of the opener code. 

The overaching principle here is that we can treat hot and cold storages separately and just open/create each and only combine them into NodeStorage as the last step. 

Some side effects of this refactoring are:
- That the opener will apply migrations to both store independently. Previously only the hot storage had migrations applied to it. It's not what we initially discussed but I think it doesn't hurt. 
- We don't need to do as many check on the kind and version - instead we just independently make sure that both hot and cold have correct versions and kinds. 
- The error messages are a bit more precise - previously we were always logging path() which only uses the hot storage, now we use path from the correct temperature storage. This unfortunately is only in the opener itself but could be extended beyond.