fix(mirror): don't convert SignedTransactions to SignedTransactionViews (#8530)
when --online-source is given, we start a full node for the source chain and get txs and receipts from a ViewClient. This is fine, but it actually results in incorrect DeployContract transactions, because the conversion from a SignedTransaction to a SignedTransactionView silently discards the contract code and replaces it with the code's hash. (This feels sort of sad, as it means the ViewClient is not quite complete?) So we can change the get_txs() function to return an `enum SourceTransaction`, to leave them untouched if we're running without --online-source. That way we can at least send DeployContract txs when --online-source is not given.