feat: type guarantees for NonDelegateAction (#8538)
Reviewer note: You may want to review this commit by commit,
as I also moved code between modules.

Background: I first added an integration test for
`DelegateActionCantContainNestedOne`.
It had the node panic with an "InvalidChunk" message.

So I started thinking. Since we define recursive delegate actions
to be syntactically wrong (parser error) it really is an invalid chunk
when it includes such a transaction. I could only construct it for the
test while bypassing borsh deserialization. Such a test doesn't really
make sense, because then I violate type invariants and in such a case
it's entirely possible that the node crashes.

My conclusion is, we must forbid the creation of an invalid
`NonDelegateAction`. I changed the code to make it obvious that
there is no violating code, using private fields and only checked
constructors.

At last, I can remove `DelegateActionCantContainNestedOne` as a
specific error and instead treat this as a proper borsh deserialization
error.