Add a flag to neard configuring reconnection to reliable peers on startup (#8576)
This is a follow-up to https://github.com/near/nearcore/pull/8318, in which reliable peers are persisted to storage in the ConnectionStore.

A new boolean flag `--connect-to-reliable-peers-on-startup` is added to neard. If set to `true`, the node will attempt to reconnect to known reliable peers from storage upon starting up. The default value is `true`.

Note that setting the flag to `false` skips reconnection, but does not purge the reliable peers in storage.

Closes https://github.com/near/nearcore/issues/8580