Update borsh dependency; again (#8608)
This is re-application of commit b8196b1aa with the same subject alas
this time borsh is updated to 0.10.2.  Previous attempt to update the
dependency was reverted becauses of potential DOS which was addressed
in borsh 0.10.2 (see <https://github.com/near/borsh-rs/pull/129>).

Like before, the update is mostly a matter of replacing deserialize
function in BorshDeserialize implementations by deserialize_reader.
In a few of existing custom implementations for (de)serialization
could be removed altogether.