[TestLoop] (5/n) Add futures support to TestLoop. (#8630)
See the comment on top of test_loop/futures.rs for what this does.

We need this because we have code like https://github.com/near/nearcore/blob/1.32.0/chain/client/src/sync/state.rs#L607 that spawns a future from an Actix handler, in order to send a message to another Actor and awaiting on the response.

So for testing, this code would be refactored so that instead of `near_performance_metrics::actix::spawn`, we would use `future_spawner.spawn`, where the `future_spawner: &dyn FutureSpawner` is passed in. It's passed in either as `ActixFutureSpawner` (from the ClientActor), or as a `DelaySender<Arc<TestLoopTask>>` from a test loop handler.

Inside, when it says `.send_async(...).then(...)`, the TestLoop will provide that send_async implementation with a `DelaySender<MessageExpectingResponse<PeerManagerMessageRequest, PeerManagerMessageResponse>>`.

Both of these will go into the test loop, where the `DriveFutures` handler will drive the future that was spawned, and the handler for `MessageExpectingResponse<PeerManagerMessageRequest, PeerManagerMessageResponse>` will execute the network adapter part of the logic. In other words, with this PR we can now, in theory, directly support the testing of all actix-based code in nearcore, as long as they use Sender and AsyncSender (whose migration was already partially done).