[TestLoop] (7/n) Provide Clock from the TestLoop API. (#8638)
This adds `TestLoopBuilder::clock()` which is consistent with the virtual time maintained by the TestLoop.

This also refactors LoopEventHandler, so that instead of implementing it as a trait, it can be created by supplying a function. This allows test loop internal code to maintain the init() functionality so that the DelaySender and the Clock can be cached and provided on each handle() call. It also makes the interface simpler to use.

Also a minor improvement to provide the future_spawner() from the TestLoopBuilder as well.